Move Frameworksdependency from apple to cxx
Although darwin style frameworks are an apple concept, they are
used by the linker and compiler at a cxx level. This will allow
classes like Cxx description enhancer to know about framworks
without a cyclic dependency on apple things.
