+++ src/com/facebook/buck/apple/PrebuiltAppleFrameworkDescription.java
@@ -16,6 +16,7 @@
 package com.facebook.buck.apple;
 
 import com.facebook.buck.cxx.CxxFlags;
+import com.facebook.buck.cxx.FrameworkDependencies;
 import com.facebook.buck.model.BuildTarget;
 import com.facebook.buck.parser.NoSuchBuildTargetException;
 import com.facebook.buck.rules.AbstractDescriptionArg;
